+++ docs/src/examples/components/Menu/Usages/MenuExampleToolbar.shorthand.tsx
@@ -131,7 +131,6 @@ class MenuExampleToolbarShorthand extends React.Component {
         defaultActiveIndex={0}
         items={items}
         iconOnly
-        primary
         accessibility={toolbarBehavior}
         aria-label="Compose Editor"
       />