+++ packages/react/src/themes/teams/components/Icon/svg/icons/skypeLogo.tsx
@@ -8,5 +8,4 @@ export default {
     </svg>
   ),
   styles: {},
-  exportedAs: 'skype-logo',
 } as TeamsProcessedSvgIconSpec