+++ docs/src/examples/components/Menu/Variations/MenuExampleFluid.shorthand.tsx
@@ -7,6 +7,6 @@ const items = [
   { key: 'events', content: 'Upcoming Events' },
 ]
 
-const MenuExampleFluid = () => <Menu defaultActiveIndex={0} items={items} vertical fluid primary />
+const MenuExampleFluid = () => <Menu defaultActiveIndex={0} items={items} vertical fluid />
 
 export default MenuExampleFluid