+++ docs/src/components/ComponentDoc/ExampleSection.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Flex, Grid, Header } from '@stardust-ui/react'
+import { Grid, Header } from '@stardust-ui/react'
 
 import { Extendable } from 'src/types'
 
@@ -14,19 +14,15 @@ export type ExampleSectionProps = Extendable<{
   title: string
 }>
 
-const ExampleSection: React.FC<ExampleSectionProps> = ({ title, children, ...restProps }) => (
-  <Flex column>
-    <Flex.Item>
-      <>
-        <Header as="h2" styles={headerStyle} className="no-anchor">
-          {title}
-        </Header>
-        <Grid variables={{ gridGap: '2rem' }} columns="1">
-          {children}
-        </Grid>
-      </>
-    </Flex.Item>
-  </Flex>
+const ExampleSection: React.FC<ExampleSectionProps> = ({ title, children }) => (
+  <>
+    <Header as="h2" styles={headerStyle} className="no-anchor">
+      {title}
+    </Header>
+    <Grid variables={{ gridGap: '2rem' }} columns="1">
+      {children}
+    </Grid>
+  </>
 )
 
 export default ExampleSection