+++ docs/src/examples/components/Chat/index.tsx
@@ -1,12 +1,14 @@
 import * as React from 'react'
 
 import Content from './Content'
+import Rtl from './Rtl'
 import Types from './Types'
 
 const ChatExamples = () => (
   <>
     <Types />
     <Content />
+    <Rtl />
   </>
 )